--- api-server/index.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const { db } = require('./src/models');
-const server = require('./src/server.js');
-
-db.sync().then(() => {
-  server.start(3000);
-});
