--- api-server/src/models/food/model.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const foodModel = (sequelize, DataTypes) => sequelize.define('Food', {
-  name: { type: DataTypes.STRING, required: true },
-  calories: { type: DataTypes.NUMBER, required: true },
-  type: { type: DataTypes.ENUM('fruit', 'vegetable', 'protein'), required: true }
-});
-
-module.exports = foodModel;
