--- auth-server/src/auth/middleware/acl.js
@@ -1,20 +1,0 @@
-'use strict';
-
-module.exports = (capability) => {
-
-  return (req, res, next) => {
-
-    try {
-      if (req.user.capabilities.includes(capability)) {
-        next();
-      }
-      else {
-        next('Access Denied');
-      }
-    } catch (e) {
-      next('Invalid Login');
-    }
-
-  }
-
-}
