--- auth-server/src/auth/models/index.js
@@ -1,13 +1,0 @@
-'use strict';
-
-const userModel = require('./users.js');
-const { Sequelize, DataTypes } = require('sequelize');
-
-const DATABASE_URL = process.env.DATABASE_URL || 'sqlite:memory;';
-
-const sequelize = new Sequelize(DATABASE_URL);
-
-module.exports = {
-  db: sequelize,
-  users: userModel(sequelize, DataTypes),
-}
