+++ src/auth/middleware/acl.js
@@ -15,6 +15,6 @@ module.exports = (capability) => {
       next('Invalid Login');
     }
 
-  },
-
+  };
 };
+
