+++ src/auth/middleware/basic.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const base64 = require('base-64');
-const { users } = require('../models');
+const { users } = require('../models/index');
 
 module.exports = async (req, res, next) => {
 
@@ -11,14 +11,14 @@ module.exports = async (req, res, next) => {
   let [user, pass] = base64.decode(basic).split(':');
 
   try {
-    req.user = await users.authenticateBasic(user, pass)
+    req.user = await users.authenticateBasic(user, pass);
     next();
   } catch (e) {
-    _authError()
+    _authError();
   }
 
   function _authError() {
     res.status(403).send('Invalid Login');
   }
 
-}
+};
