+++ src/auth/models/clothes.js
@@ -3,7 +3,7 @@
 const clothesModel = (sequelize, DataTypes) => sequelize.define('Clothes', {
   name: { type: DataTypes.STRING, required: true },
   color: { type: DataTypes.STRING, required: true },
-  size: { type: DataTypes.STRING, required: true }
+  size: { type: DataTypes.STRING, required: true },
 });
 
 module.exports = clothesModel;
