+++ src/auth/models/index.js
@@ -1,13 +1,20 @@
 'use strict';
 
-const userModel = require('./users.js');
 const { Sequelize, DataTypes } = require('sequelize');
+const clothesModel = require('./clothes');
+const foodModel = require('./food');
+const Collection = require('./data-collection');
+const userSchema = require('./users');
 
-const DATABASE_URL = process.env.DATABASE_URL || 'sqlite:memory;';
+const DATABASE_URL = process.env.DATABASE_URL || 'sqlite::memory';
 
 const sequelize = new Sequelize(DATABASE_URL);
+const food = foodModel(sequelize, DataTypes);
+const clothes = clothesModel(sequelize, DataTypes);
 
 module.exports = {
   db: sequelize,
-  users: userModel(sequelize, DataTypes),
+  food: new Collection(food),
+  clothes: new Collection(clothes),
+  users: userSchema(sequelize, DataTypes),
 };
