+++ src/auth/models/users.js
@@ -53,7 +53,7 @@ const userModel = (sequelize, DataTypes) => {
       if (user) { return user; }
       throw new Error('User Not Found');
     } catch (e) {
-      throw new Error(e.message);
+      throw new Error('THis is line 56 usermdl', e.message);
     }
   };
