+++ src/auth/routes/v1.js
@@ -2,6 +2,8 @@
 
 const express = require('express');
 const dataModules = require('../models');
+const bearerAuth = require('../middleware/bearer');
+const authorization = require('../middleware/acl');
 
 const router = express.Router();
 
@@ -15,11 +17,11 @@ router.param('model', (req, res, next) => {
   }
 });
 
-router.get('/:model', handleGetAll);
-router.get('/:model/:id', handleGetOne);
-router.post('/:model', handleCreate);
-router.put('/:model/:id', handleUpdate);
-router.delete('/:model/:id', handleDelete);
+router.get('/:model', bearerAuth, authorization('read'), handleGetAll);
+router.get('/:model/:id', bearerAuth, authorization('read'), handleGetOne);
+router.post('/:model', bearerAuth, authorization('create'), handleCreate);
+router.put('/:model/:id', bearerAuth, authorization('update'), handleUpdate);
+router.delete('/:model/:id', bearerAuth, authorization('delete'), handleDelete);
 
 async function handleGetAll(req, res) {
   let allRecords = await req.model.get();
@@ -28,7 +30,7 @@ async function handleGetAll(req, res) {
 
 async function handleGetOne(req, res) {
   const id = req.params.id;
-  let theRecord = await req.model.get(id)
+  let theRecord = await req.model.get(id);
   res.status(200).json(theRecord);
 }
 
@@ -41,7 +43,7 @@ async function handleCreate(req, res) {
 async function handleUpdate(req, res) {
   const id = req.params.id;
   const obj = req.body;
-  let updatedRecord = await req.model.update(id, obj)
+  let updatedRecord = await req.model.update(id, obj);
   res.status(200).json(updatedRecord);
 }
